drasl

+++ src/main/java/artyfartyparty/solowebservice/Model/User.java
@@ -12,7 +12,7 @@ public class User {
 
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
     private String name;
     private String uniMail;
     private String Address;
@@ -30,12 +30,11 @@ public class User {
         this.name = name;
     }
 
-    public Integer getId() {
-
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
